added transition between menu and survey questions
